* Fixes #2914: Fixes file descriptor confusion on 64bit systems. Reviewed by cgibbons.
git-svn-id: http://dev.zenoss.com/svnint/trunk/core/pynetsnmp@25847 f5df01ce-0320-0410-9ed9-ddcd13c732cc

+++ netsnmp.py
@@ -510,7 +510,7 @@ class Session(object):
     
 
 MAXFD = 1024
-fdset = c_long * (MAXFD/32)
+fdset = c_int32 * (MAXFD/32)
 
 class timeval(Structure):
     _fields_ = [
